Community Grid Model Class

Community Grid Model Class

Community Grid Model Class

Details

A community_model whose pattern is a regular, rectangular grid of points.

Super class

SimComm::community_model -> community_gridmodel

Public fields

tess

A tessellation of the window containing the points, used to plot the model.

neighborhood

A character string defining what is the neighborhood of a cell: "von Neumann 1" or "4" for the closest four neighbors (North, West, South, East); "Moore 1" or "8" for all adjacent cells (the first four and North-West, etc.); "Moore 2" or "24" for two rings of neighbors.

Methods

Public methods

community_gridmodel$new()

community_gridmodel$neighbors()

community_gridmodel$plot()

community_gridmodel$clone()

Inherited methods

SimComm::community_model$along_time()

SimComm::community_model$autoplot()

SimComm::community_model$run()

SimComm::community_model$saved_pattern()

Method new()

Create a new instance of this R6 class.

Usage

community_gridmodel$new(
  pattern = pattern_grid(),
  timeline = 0,
  type = "Species",
  neighborhood = "von Neumann 1"
)

Arguments

pattern

The pattern which describes the location of agents.

timeline

A numeric vector that contains the points of time of interest.

type

The type of individuals. Informational only.

neighborhood

A character string defining what is the neighborhood of a cell: "von Neumann 1" or "4" for the closest four neighbors (North, West, South, East); "Moore 1" or "8" for all adjacent cells (the first four and North-West, etc.); "Moore 2" or "24" for two rings of neighbors.

Method neighbors()

Return a vector that contains the indices of neighbors in the point pattern.

Usage

community_gridmodel$neighbors(point)

Arguments

point

The focal point around which a neighborhood is defined.

Method plot()

Plots the pattern.

Usage

community_gridmodel$plot(
  time = NULL,
  sleep = 0,
  which.marks = "PointType",
  ...
)

Arguments

time

The point of the timeline considered. Its value should be in timeline.

sleep

The time (in ms) to sleep between each step of the evolution of the model.

which.marks

The marks to plot, that must be a the name of a column of the mark dataframe.

...

Extra arguments to be passed to methods.

Method clone()

The objects of this class are cloneable with this method.

Usage

community_gridmodel$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.