Community Drift

Community Drift

Community Drift

Details

A community_spcmodel. At each generation, each individual is replaced by one of its neighbors.

Super classes

SimComm::community_model -> SimComm::community_spcmodel -> cp_drift

Public fields

n_neighbors

The number of nearest neighbors to take into account.

Methods

Public methods

cp_drift$new()

cp_drift$clone()

Inherited methods

SimComm::community_model$along_time()

SimComm::community_model$autoplot()

SimComm::community_model$run()

SimComm::community_model$saved_pattern()

SimComm::community_spcmodel$neighbor_types_r()

SimComm::community_spcmodel$neighbors_n()

SimComm::community_spcmodel$neighbors_r()

SimComm::community_spcmodel$plot()

Method new()

Create a new object.

Usage

cp_drift$new(
  pattern = SpatDiv::rSpCommunity(n = 1, size = 100, CheckArguments = FALSE),
  timeline = 0,
  type = "Species",
  n_neighbors = 6
)

Arguments

pattern

The pattern which describes the location of agents.

timeline

A numeric vector that contains the points of time of interest.

type

The type of individuals. Informational only.

n_neighbors

Number of neighbors.

Method clone()

The objects of this class are cloneable with this method.

Usage

cp_drift$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.